Extract useConfigurationEditor from the configurations page
